-Create a new dao, -Create UserDataSource -Create Medicine Data source and refactor the repository -Create Pet Data source and refactor the repository

+++ app/src/main/java/nondas/pap/petcareapp/domain/usecase/medicine/GetMedicine.kt
@@ -21,6 +21,6 @@ class GetMedicine @Inject constructor(
     }
 
 
-    data class Params(val petId: Long)
+    data class Params(val petId: Int)
 
 }
